requirements.txt
opensmile
faster-whisper
syllables
typer
pandas
numpy
ffmpeg-python
openai-whisper
nltk
matplotlib
textstat
